First pass at the rev 2 power board.
Only did copper layers on the power board so far. Waiting to decide on a PCB fab house then I'll go back and make sure it's compliant (looking mostly at how close I can pack vias). Starting on the nucleo programmer shield next. Added a voltage divider to the enable pin on the 3V3 vreg on the controller board so it only switches on when there is enough input voltage for correct 3V3 output. I did this because I read somewhere that the RP2040 can power on at a lower voltage than a lot of flash chips and if that happens it gets into a weird state so I'm trying to make sure 3V3 comes on as quickly and sharply as possible.